+++ src/components/scenes/settings/SettingsSceneShotsDetailButton.jsx
@@ -1,7 +1,7 @@
 import ListAltIcon from '@material-symbols/svg-400/rounded/list_alt.svg';
 
 import SettingsFieldButton from '@/components/settings/SettingsFieldButton';
-import { useSetUserCursor, useUserStore } from '@/stores/user';
+import { useUserStore } from '@/stores/user';
 
 /**
  * @param {object} props
@@ -15,15 +15,12 @@ export default function SettingsSceneShotsDetailButton({
   const shotListMode = useUserStore((ctx) => ctx.shotListMode);
   const setShotListMode = useUserStore((ctx) => ctx.setShotListMode);
   const userCursor = useUserStore((ctx) => ctx.cursor);
-  const setUserCursor = useSetUserCursor();
 
   function onClick() {
     if (shotListMode !== 'detail' || userCursor.shotId !== '') {
       setShotListMode('detail');
-      setUserCursor(documentId, sceneId, '');
     } else {
       setShotListMode('hidden');
-      setUserCursor(documentId, '', '');
     }
   }