+++ src/components/shots/ShotList.jsx
@@ -1,6 +1,7 @@
 import { Fragment } from 'react';
 
-import { useShotIds } from '@/stores/document';
+import { useBlockIds, useShotIds } from '@/stores/document';
+import { useUserStore } from '@/stores/user';
 
 import TakeList from '../takes/TakeList';
 import GridStyle from './GridStyle.module.css';
@@ -13,9 +14,10 @@ import ShotEntryNew from './ShotEntryNew';
  * @param {string} [props.className]
  * @param {import('@/stores/document/DocumentStore').DocumentId} props.documentId
  * @param {import('@/stores/document/DocumentStore').SceneId} props.sceneId
- * @param {import('@/stores/document/DocumentStore').BlockId} props.blockId
+ * @param {import('@/stores/document/DocumentStore').BlockId} [props.blockId]
  * @param {boolean} [props.editable]
  * @param {boolean} [props.collapsed]
+ * @param {boolean} [props.hidden]
  */
 export default function ShotList({
   className,
@@ -24,23 +26,94 @@ export default function ShotList({
   blockId,
   editable = true,
   collapsed = false,
+  hidden = false,
+}) {
+  const hasActiveShot = useUserStore((ctx) => Boolean(ctx.cursor?.shotId));
+  const blockIds = useBlockIds(documentId, sceneId);
+  const lastBlockId = blockId || blockIds.at(-1);
+  return (
+    <ul
+      className={
+        hidden
+          ? /* NOTE: Quick hideaway to not lag. */ 'invisible absolute top-0 left-0 -z-10'
+          : className
+      }>
+      <div className={collapsed ? GridStyle.grid : ''}>
+        <PerBlock blockIds={blockId ? [blockId] : blockIds}>
+          {(blockId) => (
+            <ShotsByBlock
+              documentId={documentId}
+              sceneId={sceneId}
+              blockId={blockId}
+              editable={editable}
+              collapsed={collapsed}
+              showTakes={hasActiveShot}
+              showNew={editable && lastBlockId === blockId}
+            />
+          )}
+        </PerBlock>
+      </div>
+      <ShotEntryDragged documentId={documentId} sceneId={sceneId} />
+    </ul>
+  );
+}
+
+/**
+ * @param {object} props
+ * @param {Array<import('@/stores/document/DocumentStore').BlockId>} props.blockIds
+ * @param {(
+ * blockId: import('@/stores/document/DocumentStore').BlockId,
+ * index: number,
+ * array: Array<import('@/stores/document/DocumentStore').BlockId>
+ * ) => import('react').ReactNode} props.children
+ */
+function PerBlock({ blockIds, children }) {
+  return (
+    <>
+      {blockIds.map((blockId, index, array) => (
+        <Fragment key={`block-${blockId}`}>
+          {children(blockId, index, array)}
+        </Fragment>
+      ))}
+    </>
+  );
+}
+
+/**
+ * @param {object} props
+ * @param {import('@/stores/document/DocumentStore').DocumentId} props.documentId
+ * @param {import('@/stores/document/DocumentStore').SceneId} props.sceneId
+ * @param {import('@/stores/document/DocumentStore').BlockId} props.blockId
+ * @param {boolean} props.editable
+ * @param {boolean} props.collapsed
+ * @param {boolean} props.showTakes
+ * @param {boolean} props.showNew
+ */
+function ShotsByBlock({
+  documentId,
+  sceneId,
+  blockId,
+  editable,
+  collapsed,
+  showTakes,
+  showNew,
 }) {
   const shotIds = useShotIds(documentId, blockId);
   if (shotIds.length <= 0) {
     return null;
   }
   return (
-    <ul className={className}>
-      <div className={collapsed ? GridStyle.grid : ''}>
-        {shotIds.map((shotId) => (
-          <Fragment key={`shot-${shotId}`}>
-            <ShotEntry
-              documentId={documentId}
-              sceneId={sceneId}
-              blockId={blockId}
-              shotId={shotId}
-              editable={editable}
-              collapsed={collapsed}>
+    <>
+      {shotIds.map((shotId) => (
+        <Fragment key={`shot-${shotId}`}>
+          <ShotEntry
+            documentId={documentId}
+            sceneId={sceneId}
+            blockId={blockId}
+            shotId={shotId}
+            editable={editable}
+            collapsed={collapsed}>
+            {showTakes && (
               <TakeList
                 documentId={documentId}
                 sceneId={sceneId}
@@ -48,19 +121,18 @@ export default function ShotList({
                 shotId={shotId}
                 viewMode={collapsed ? 'list' : 'inline'}
               />
-            </ShotEntry>
-          </Fragment>
-        ))}
-        {editable && (
-          <ShotEntryNew
-            documentId={documentId}
-            sceneId={sceneId}
-            blockId={blockId}
-            collapsed={collapsed}
-          />
-        )}
-      </div>
-      <ShotEntryDragged documentId={documentId} sceneId={sceneId} />
-    </ul>
+            )}
+          </ShotEntry>
+        </Fragment>
+      ))}
+      {showNew && (
+        <ShotEntryNew
+          documentId={documentId}
+          sceneId={sceneId}
+          blockId={blockId}
+          collapsed={collapsed}
+        />
+      )}
+    </>
   );
 }